Add first build switches
`defines()` and `includeDirs()` are now implemented for Visual Studio, at both the project and the file level

+++ core/host/src/libraries/pmk_lib_terminal.c
@@ -8,7 +8,7 @@
 int pmk_terminal_textColor(lua_State* L)
 {
 	if (lua_gettop(L) > 0) {
-		int color = luaL_checkinteger(L, 1);
+		int color = (int)luaL_checkinteger(L, 1);
 		pmk_setTextColor(color);
 	}
